# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/repairinvoice/{id}": {
            "delete": {
                "description": "get repairinvoice by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a repairinvoice entity by ID",
                "operationId": "delete-repairinvoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repairinvoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/repairinvoices": {
            "get": {
                "description": "list repairinvoice entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List repairinvoice entities",
                "operationId": "list-repairinvoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Repairinvoice"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create repairinvoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create repairinvoice",
                "operationId": "create-repairinvoice",
                "parameters": [
                    {
                        "description": "Repairinvoice entity",
                        "name": "repairinvoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Repairinvoice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Repairinvoice"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/repairinvoices/{id}": {
            "get": {
                "description": "get repairinvoice by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a repairinvoice entity by ID",
                "operationId": "get-repairinvoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repairinvoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Repairinvoice"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/returninvoices": {
            "get": {
                "description": "list returninvoice entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List returninvoice entities",
                "operationId": "list-returninvoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Returninvoice"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create returninvoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create returninvoice",
                "operationId": "create-returninvoice",
                "parameters": [
                    {
                        "description": "Returninvoice entity",
                        "name": "returninvoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Returninvoice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Returninvoice"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Returninvoice": {
            "type": "object",
            "properties": {
                "addedtime": {
                    "type": "string"
                },
                "employee": {
                    "type": "integer"
                },
                "repairinvoice": {
                    "type": "integer"
                },
                "statust": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Employee": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EmployeeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.EmployeeEdges"
                },
                "employeeemail": {
                    "description": "Employeeemail holds the value of the \"employeeemail\" field.",
                    "type": "string"
                },
                "employeename": {
                    "description": "Employeename holds the value of the \"employeename\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.EmployeeEdges": {
            "type": "object",
            "properties": {
                "employees": {
                    "description": "Employees holds the value of the employees edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Returninvoice"
                    }
                }
            }
        },
...
        "ent.Repairinvoice": {
            "type": "object",
            "properties": {
                "deviceid": {
                    "description": "Deviceid holds the value of the \"deviceid\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RepairinvoiceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RepairinvoiceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "statusrepairid": {
                    "description": "Statusrepairid holds the value of the \"statusrepairid\" field.",
                    "type": "integer"
                },
                "symptomid": {
                    "description": "Symptomid holds the value of the \"symptomid\" field.",
                    "type": "integer"
                },
                "userid": {
                    "description": "Userid holds the value of the \"userid\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.RepairinvoiceEdges": {
            "type": "object",
            "properties": {
                "repairinvoices": {
                    "description": "Repairinvoices holds the value of the repairinvoices edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Returninvoice"
                }
            }
        },
        "ent.Returninvoice": {
            "type": "object",
            "properties": {
                "addedtime": {
                    "description": "Addedtime holds the value of the \"addedtime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ReturninvoiceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ReturninvoiceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ReturninvoiceEdges": {
            "type": "object",
            "properties": {
                "employee": {
                    "description": "Employee holds the value of the Employee edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Employee"
                },
                "repairinvoice": {
                    "description": "Repairinvoice holds the value of the Repairinvoice edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Repairinvoice"
                },
                "statust": {
                    "description": "Statust holds the value of the Statust edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Statust"
                }
            }
        },
...
        "ent.Statust": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StatustQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StatustEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "statustname": {
                    "description": "Statustname holds the value of the \"statustname\" field.",
                    "type": "string"
                }
            }
        },
        "ent.StatustEdges": {
            "type": "object",
            "properties": {
                "statusts": {
                    "description": "Statusts holds the value of the statusts edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Returninvoice"
                    }
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}